test(py/plugins/openai): move the gpt-6-astra json_object test to the model tests
It builds an OpenAIModel, not the plugin, so it belongs next to the
unlisted-id case it complements.
