+++ py/packages/genkit-openai/tests/openai_model_test.py
@@ -54,6 +54,16 @@ def test_unknown_chat_id_json_mode_uses_json_object() -> None:
     assert model._get_response_format(request) == {'type': 'json_object'}
 
 
+def test_gpt_6_astra_json_mode_uses_json_object() -> None:
+    """A schema-less JSON request to gpt-6-astra sends json_object, as the catalog advertises."""
+    model = OpenAIModel(model='gpt-6-astra', client=MagicMock())
+    request = ModelRequest(
+        messages=[Message(role=Role.USER, content=[Part(root=TextPart(text='Hi'))])],
+        output=OutputConfig(format='json'),
+    )
+    assert model._get_response_format(request) == {'type': 'json_object'}
+
+
 def test_get_messages(sample_request: ModelRequest) -> None:
     """Test _get_messages method.
