+++ py/packages/genkit-openai/tests/openai_plugin_test.py
@@ -24,14 +24,12 @@ from typing import Any, cast
 from unittest.mock import AsyncMock, MagicMock
 
 import pytest
-from genkit_openai.models import OpenAIModel
 from genkit_openai.models.model_info import SUPPORTED_OPENAI_MODELS
 from genkit_openai.openai_plugin import OpenAI, openai_model
 from genkit_openai.typing import SupportedOutputFormat
 from openai.types import Model
 
-from genkit import Message, ModelRequest, Part, Role, Supports, TextPart
-from genkit._core._model import OutputConfig
+from genkit import Supports
 from genkit.plugin_api import ActionKind, ActionMetadata, loop_local_client
 
 
@@ -133,16 +131,6 @@ async def test_unlisted_chat_model_resolves_with_default_supports() -> None:
     assert action.metadata is not None
     model_meta = cast(dict[str, Any], action.metadata['model'])
     assert model_meta['supports'] == {'multiturn': True}
-
-
-def test_gpt_6_astra_json_mode_uses_json_object() -> None:
-    """A schema-less JSON request to gpt-6-astra sends json_object, as the catalog advertises."""
-    model = OpenAIModel(model='gpt-6-astra', client=MagicMock())
-    request = ModelRequest(
-        messages=[Message(role=Role.USER, content=[Part(root=TextPart(text='Hi'))])],
-        output=OutputConfig(format='json'),
-    )
-    assert model._get_response_format(request) == {'type': 'json_object'}
 
 
 @pytest.mark.asyncio
